+++ hazelcastmq-examples/src/main/java/org/mpilone/hazelcastmq/example/stomp/StompWithExternalClient.java
@@ -35,7 +35,9 @@ public class StompWithExternalClient {
 
   public static void main(String[] args) throws Exception {
     System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "debug");
+    System.setProperty("org.slf4j.simpleLogger.showDateTime", "true");
     System.setProperty("org.slf4j.simpleLogger.log.com.hazelcast", "info");
+    System.setProperty("org.slf4j.simpleLogger.log.io.netty", "info");
 
     new StompWithExternalClient();
   }